Fixed Some Stuff on Cam

+++ satellite-subsystems/demo/src/sub-systemCode/payload/DataBase/DataBase.c
@@ -404,7 +404,7 @@ ImageDataBaseResult transferImageToSD_withoutSearch(imageid cameraId, uint32_t i
 	error = checkForFileType(image_metadata, raw);
 	CMP_AND_RETURN(error, DataBaseSuccess, DataBaseFail);
 
-	WritePayloadLog(PAYLOAD_TRANSFERRED_IMAGE, image_metadata.cameraId);
+	WritePayloadLog(PAYLOAD_TRANSFERRED_IMAGE, (uint32_t)image_metadata.cameraId);
 
 	return DataBaseSuccess;
 }
@@ -500,7 +500,7 @@ ImageDataBaseResult DeleteImageFromPayload(ImageDataBase database, imageid id)
 
 	updateGeneralDataBaseParameters(database);
 
-	WritePayloadLog(PAYLOAD_ERASED_IMAGE, image_metadata.cameraId);
+	WritePayloadLog(PAYLOAD_ERASED_IMAGE, (uint32_t)image_metadata.cameraId);
 
 	return DataBaseSuccess;
 }
@@ -669,7 +669,7 @@ ImageDataBaseResult takePicture(ImageDataBase database, Boolean8bit testPattern)
 
 	updateGeneralDataBaseParameters(database);
 
-	WritePayloadLog(PAYLOAD_TOOK_IMAGE, getLatestID(database));
+	WritePayloadLog(PAYLOAD_TOOK_IMAGE, (uint32_t)getLatestID(database));
 
 	return DataBaseSuccess;
 }
